% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.R
\name{get_comments}
\alias{get_comments}
\title{Return media and media metadata}
\usage{
get_comments(
  username = NULL,
  tag = NULL,
  location = NULL,
  your_username = NULL,
  your_password = NULL,
  following_input = FALSE,
  max = NULL,
  include_location = FALSE,
  interactive = NULL,
  filter = NULL,
  filter_location = NULL,
  raw_data = TRUE,
  retry_download = FALSE
)
}
\arguments{
\item{username}{String vector. Specify a username account}

\item{tag}{String. Scrapes the specified hashtag for media.}

\item{location}{String. Scrapes the specified instagram location-id for media.}

\item{your_username}{String. Instagram login user}

\item{your_password}{String. Instagram login password}

\item{following_input}{Logical value.  Use profiles followed by login-user as input}

\item{max}{Maximum number of items to scrape.}

\item{include_location}{Logical value. Includes location metadata when saving media metadata.}

\item{interactive}{Enables interactive login challenge solving. Has 2 modes: SMS and Email}

\item{filter}{String vector.  Scrapes the specified hashtag within a user's media.}

\item{filter_location}{String vector.Filter scrape queries by command line location(s) ids}

\item{raw_data}{Logical value. if TRUE, clean metadata.}

\item{retry_download}{Logical values. Retry download attempts endlessly when errors are received}
}
\value{
A tibble and profile metadata
}
\description{
\code{get_media} return media and media metadata based on search configuration.
}
\examples{
\dontrun{
get_media(following_input = TRUE, your_username = "xxxx", your_password = "123")
get_media(username = "ABC", max = 10, media = "image")
}

}
