% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profiles.R
\name{get_profiles}
\alias{get_profiles}
\title{Return user profile informations}
\usage{
get_profiles(
  username = NULL,
  your_username = NULL,
  your_password = NULL,
  following_input = FALSE,
  retry_download = FALSE
)
}
\arguments{
\item{username}{String vector. Specify  username accounts}

\item{your_username}{String. Instagram login user}

\item{your_password}{String. Instagram login password}

\item{following_input}{Logical value.  Use profiles followed by login-user as input}

\item{retry_download}{Logical values. Retry download attempts endlessly when errors are received}
}
\value{
A tibble and profile metadata
}
\description{
\code{get_profiles} return profile informations from username input.
}
\examples{
\dontrun{
get_profiles(following_input = TRUE, your_username = "xxxx", your_password = "123")
}

}
