% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instascraper_install.R
\name{instascraper_install}
\alias{instascraper_install}
\title{Set up instagram scraper}
\usage{
instascraper_install()
}
\description{
\code{instascraper_install} will install instagram scraper from pip.
}
\examples{
\dontrun{
instascraper_install()
}

}
