% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{parse_json}
\alias{parse_json}
\title{Parse a json file}
\usage{
parse_json(path = NULL, type = NULL, input_argument = NULL)
}
\arguments{
\item{path}{String that specify file location}

\item{type}{String. Specify the parsing type. Possible values: comments (a json file comes from \code{get_comments}), media (a json file comes from \code{get_media}), or profiles (a json file comes from \code{get_profiles}).}

\item{input_argument}{String. Specify the first argument you use in function \code{get_comments}, \code{get_media}, and \code{get_profiles}. Possible values: username, tag, location, or none.}
}
\value{
A tibble
}
\description{
\code{parse_json} parse a json file based on specific configuration.
}
\examples{
\dontrun{
parse_json("x/x.json", type = "profiles", input_argument = "none")
}

}
