% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_location.R
\name{search_location}
\alias{search_location}
\title{Search location id}
\usage{
search_location(name = NULL)
}
\arguments{
\item{name}{Character that specify a location name .}
}
\value{
A sting
}
\description{
\code{search_location} returns location id based on location name.
}
\examples{
\dontrun{
search_location("Bandung")
}

}
